+++ include/mep_stats.h
@@ -27,18 +27,18 @@ public:
 
 	// could be less than the number of generations of the algorithm if the user stops it earlier
 	int last_generation;
-	int num_generations;
+	unsigned int num_generations;
 
 	t_mep_run_statistics(void);
 	~t_mep_run_statistics();
 
-	void allocate_memory(long num_generations);
+	void allocate_memory(unsigned int num_generations);
 	void delete_memory(void);
 
 	t_mep_run_statistics& operator=(const t_mep_run_statistics& source);
 
 	int to_xml(pugi::xml_node parent);
-	int from_xml(pugi::xml_node parent, int problem_type);
+	int from_xml(pugi::xml_node parent, unsigned int problem_type);
 
 };
 //-----------------------------------------------------------------
@@ -56,13 +56,13 @@ private:
 
 public:
 
-	int num_runs;
+	unsigned int num_runs;
 
 	t_mep_statistics();
 
 	t_mep_statistics& operator=(const t_mep_statistics& source);
 
-	t_mep_run_statistics *get_stat_ptr(int index)  const;
+	t_mep_run_statistics *get_stat_ptr(unsigned int index)  const;
 
 	void get_best_error(double &training, double &validation, double &test, double &running_time) const;
 	void get_mean_error(double &training, double &validation, double &test, double &running_time) const;
@@ -73,38 +73,38 @@ public:
 	void get_sttdev_num_incorrect(double &training, double &validation, double &test, double &running_time) const;
 
 	// returns the best training error
-	double get_best_training_error(int run, int generation) const;
+	double get_best_training_error(unsigned int run, unsigned int generation) const;
 
 	// returns the best validation error
-	double get_best_validation_error(int run) const;
+	double get_best_validation_error(unsigned int run) const;
 
 	// returns the average (over the entire population) training error
-	double get_average_training_error(int run, int generation) const;
+	double get_average_training_error(unsigned int run, unsigned int generation) const;
 
 	// returns the best training num incorreclty classified
-	double get_best_training_num_incorrectly_classified(int run, int generation) const;
+	double get_best_training_num_incorrectly_classified(unsigned int run, unsigned int generation) const;
 
 	// returns the best validation num incorreclty classified
-	double get_best_validation_num_incorrectly_classified(int run) const;
+	double get_best_validation_num_incorrectly_classified(unsigned int run) const;
 
 	// returns the average (over the entire population) training num incorreclty classified
-	double get_mean_training_num_incorrectly_classified(int run, int generation) const;
+	double get_mean_training_num_incorrectly_classified(unsigned int run, unsigned int generation) const;
 
 	// returns the average (over the entire population) training num incorreclty classified
-	double get_stddev_training_num_incorrectly_classified(int run, int generation);
+	double get_stddev_training_num_incorrectly_classified(unsigned int run, unsigned int generation);
 
 	// returns the running time
-	double get_running_time(int run) const;
+	double get_running_time(unsigned int run) const;
 
 	// returns the error on the test data
-	double get_test_error(int run) const;
+	double get_test_error(unsigned int run) const;
 
 	// returns the num incorreclty classified on the test data
-	double get_test_num_incorrectly_classified(int run) const;
+	double get_test_num_incorrectly_classified(unsigned int run) const;
 
 	// save statistics to csv file
-	int to_csv(const char* file_name, int problem_type)const;
-	int to_tex(const char* file_name, int problem_type)const;
+	int to_csv(const char* file_name, unsigned int problem_type)const;
+	int to_tex(const char* file_name, unsigned int problem_type)const;
 
 	void sort_stats_by_running_time(bool ascending);
 	void sort_stats_by_training_error(bool ascending);
@@ -112,16 +112,16 @@ public:
 	void sort_stats_by_test_error(bool ascending);
 
 	// returns the last generation of a given run (useful when the run has been stopped earlier)
-	int get_latest_generation(int run) const;
+	int get_latest_generation(unsigned int run) const;
 
 	void delete_memory(void);
-	void append(int num_generations);
-	void create(int _num_runs);
+	void append(unsigned int num_generations);
+	void create(unsigned int _num_runs);
 
-	void compute_mean_stddev(bool compute_on_validation, bool compute_on_test, int problem_type);
+	void compute_mean_stddev(bool compute_on_validation, bool compute_on_test, unsigned int problem_type);
 
 	int to_xml(pugi::xml_node parent);
-	int from_xml(pugi::xml_node parent, int problem_type);
+	int from_xml(pugi::xml_node parent, unsigned int problem_type);
 };
 //-----------------------------------------------------------------
 #endif
